Реализован метод создания дружбы

+++ src/main/java/ru/yandex/practicum/filmorate/exception/GenreNotFoundException.java
@@ -1,7 +1,7 @@
 package ru.yandex.practicum.filmorate.exception;
 
 public class GenreNotFoundException extends NotFoundException{
-    public GenreNotFoundException(Long id) {
+    public GenreNotFoundException(int id) {
         super(id);
         message = "Жанр с id=%d не найден".formatted(id);
     }
